backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
google-generativeai==0.3.2
pydantic==2.5.2
python-multipart==0.0.6 